feat: add multimodal for vertex ai anthropic
